requirements.txt
Django>=1.8,<2.0
djangorestframework
django-crispy-forms
django-filter
django-request
django-guardian
Markdown
Pigments
coreapi
psycopg2
